apps/tracker/pyproject.toml
[project]
name = "ai-email-tracker"
version = "2.0.0"
description = "AI-powered Job Application Email Tracker — replaces n8n workflow with a code-first Python pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Maher", email = "maherahmedraza1@gmail.com" }]

dependencies = [
    "google-api-python-client>=2.130.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-genai>=1.0.0",
    "supabase>=2.6.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "thefuzz[speedup]>=0.22.1",
    "loguru>=0.7.2",
    "requests>=2.31.0",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["*.py"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
